// src/wca-api/openapiClient/types.gen.ts
export type WcifTimeLimit = {
    centiseconds: number;
    cumulativeRoundIds: Array<string>;
};

export type WcifCutoff = {
    numberOfAttempts: number;
    attemptResult: WcifAttemptResult;
};

export type WcifAdvancementConditionRanking = {
    type: string;
    level: WcifRanking;
};

export type WcifAdvancementConditionPercent = {
    type: string;
    level: WcifPercent;
};

export type WcifAdvancementConditionAttemptResult = {
    type: string;
    level: WcifAttemptResult;
};

export type WcifAdvancementCondition = ({
    type: 'ranking';
} & WcifAdvancementConditionRanking) | ({
    type: 'percent';
} & WcifAdvancementConditionPercent) | ({
    type: 'attemptResult';
} & WcifAdvancementConditionAttemptResult);

export type WcifScramble = string;

export type WcifScrambleSet = {
    id: number;
    scrambles: Array<WcifScramble>;
    extraScrambles: Array<WcifScramble>;
};

export type BaseWcifRound = {
    id: string;
    format: string;
    timeLimit?: WcifTimeLimit;
    cutoff?: WcifCutoff;
    advancementCondition?: WcifAdvancementCondition;
    scrambleSetCount: number;
    scrambleSets: Array<WcifScrambleSet>;
    extensions: Array<unknown>;
};

export type LiveAttempt = {
    value: number;
    attempt_number: number;
};

export type BaseLiveResult = {
    registration_id: number;
    best: number;
    average: number;
    single_record_tag: string;
    average_record_tag: string;
    advancing: boolean;
    advancing_questionable: boolean;
    attempts: Array<LiveAttempt>;
    last_attempt_entered_at: string;
};

export type RoundLiveResult = BaseLiveResult & {
    round_wcif_id: string;
    global_pos: number;
    local_pos: number;
};

export type LiveCompetitor = {
    id: number;
    registrant_id: number;
    user_id: number;
    name: string;
    country_iso2: string;
};

export type LiveRound = BaseWcifRound & {
    results: Array<RoundLiveResult>;
    competitors: Array<LiveCompetitor>;
    round_id: number;
    state_hash: string;
    linked_round_ids?: Array<string>;
};

export type Result = {
    id: number;
    pos: number;
    best: number;
    average: number;
    name: string;
    country_iso2: string;
    competition_id: string;
    event_id: string;
    round_type_id: string;
    format_id: string;
    wca_id: string;
    attempts: Array<number>;
    best_index: number;
    worst_index: number;
    regional_single_record: string;
    regional_average_record: string;
};